make the number of seconds values and number of records value as Long Integer.

+++ driver-foundationdb/src/main/java/io/sbk/FoundationDB/FoundationDBMultiKeyWriter.java
@@ -71,7 +71,7 @@ public class FoundationDBMultiKeyWriter implements Writer<byte[]> {
     public void writeAsyncTime(DataType<byte[]> dType, byte[] data, int size, Time time, Status status) throws IOException {
         final int recs;
         if (params.getRecordsPerWriter() > 0 && params.getRecordsPerWriter() > cnt) {
-            recs = Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
+            recs = (int) Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
         } else {
             recs = params.getRecordsPerSync();
         }
@@ -95,7 +95,7 @@ public class FoundationDBMultiKeyWriter implements Writer<byte[]> {
                             Status status, SendChannel sendChannel, int id) throws IOException {
         final int recs;
         if (params.getRecordsPerWriter() > 0 && params.getRecordsPerWriter() > cnt) {
-            recs = Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
+            recs = (int) Math.min(params.getRecordsPerWriter() - cnt, params.getRecordsPerSync());
         } else {
             recs = params.getRecordsPerSync();
         }
